-D bug fixing; 64bit bootstrap
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/lw-ipo@150675 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/l-ipo.c
@@ -485,7 +485,7 @@ type_addr_hash (const void *ent)
 {
   const struct type_ent *const entry
       = (const struct type_ent *) ent;
-  return (hashval_t) entry->type;
+  return (hashval_t) (long) entry->type;
 }
 
 /* Address equality function for type_ent.  */
